Just got removeIndex in SinglylinkedList to work. Awful.
It took me like two hours to do this, because even though I got the
basic concept, it was the stupid little things. I was at first just
setting the nodes equal to the next ones instead of just the values, and
then it wouldn't return the proper value that it was at for whatever
reason. Awful.